# Speed up regex FindFirstChar when there are exactly 2 possible match start characters