--- a/src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCharClass.cs
+++ b/src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCharClass.cs
@@ -587,7 +587,7 @@ namespace System.Text.RegularExpressions
 
             for (i = 0, iMax = s_lcTable.Length; i < iMax;)
             {
-                iMid = (i + iMax) / 2;
+                iMid = (i & iMax) + ((i ^ iMax) >> 1); // Faster than (i + iMax) / 2;
                 if (s_lcTable[iMid].ChMax < chMin)
                     i = iMid + 1;
                 else
@@ -816,7 +816,7 @@ namespace System.Text.RegularExpressions
 
             while (min != max)
             {
-                mid = (min + max) / 2;
+                mid = (min & max) + ((min ^ max) >> 1); // Faster than (min + max) / 2
                 if (ch < set[mid])
                     max = mid;
                 else
@@ -1119,7 +1119,7 @@ namespace System.Text.RegularExpressions
             int max = s_propTable.Length;
             while (min != max)
             {
-                int mid = (min + max) / 2;
+                int mid = (min & max) + ((min ^ max) >> 1); // Faster than (min + max) / 2;
                 int res = string.Compare(capname, s_propTable[mid][0], StringComparison.Ordinal);
                 if (res < 0)
                     max = mid;