Regex-Redux spends much of its time scanning its huge input string to find the first character of a match ([see perf analysis here](https://github.com/dotnet/corefx/issues/27124) - this is the 20% or more time noted in CharInClassInternal)

This scan is the job of FindFirstChar(). In general character classes are complicated objects because they must store possibly several valid ranges, which may be character classes, as well as some options such as negating. They are packed objects and a binary search is used to check whether a candidate character matches the class. See http://www.moserware.com/2009/03/how-net-regular-expressions-really-work.html for a breakdown of how they work.

We do some optimizations to avoid this binary search for every character. First is Boyer-Moore which helps for long fixed prefix. Second is [IsSingleton](https://github.com/dotnet/corefx/blob/master/src/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexInterpreter.cs#L420), which special cases for positive match of a single specific character.

RegexRedux doesn't benefit much from either of these. It has no long prefix and in many cases there are several possible first characters. For example [here](https://github.com/dotnet/corefx/blob/master/src/System.Text.RegularExpressions/tests/Performance/Perf.RegexRedux.cs#L56) in `"agggt[cgt]aa|tt[acg]accct"` much of the time it is scanning for either `a` or `t`. In some cases there are some more options for example in `"agggtaa[cgt]|[acg]ttaccct"` it would start looking for `a`, `c` or `g` to begin the match. However two characters is the key case.

At this point we are looking for localized low risk changes.

Firstly I made a small speedup to the binary search by replacing `(min + max) / 2` with ` (min & max) + ((min ^ max) >> 1)` (apparently from Knuth .. it can be intuitively seen that halving the sum will retain all the bits they have in common, and half of all the bits they do not have in common). It is nice that it avoids overflow but that is not relevant in this case. This gave only a small improvement: I measured ~1% on an artificial test case and within noise on the full test. Nevertheless it is low harm so I retained it and used it in the other two binary searches in the file.

Secondly I special cased the "two possible characters" as IsDoubleton in the same pattern as existing IsSingleton. This gave a larger benefit, with no measureable hit on the generic test case:

## before
two runs:

  System.Text.RegularExpressions.Performance.Tests.dll                               | Metric         | Unit  | Iterations |    Average |   STDEV.S |        Min |        Max
  :--------------------------------------------------------------------------------- |:-------------- |:-----:|:----------:| ----------:| ---------:| ----------:| ----------:
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | Duration       | msec  |     44     |    229.951 |    11.146 |    219.932 |    258.559
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | GC Allocations | bytes |     44     | 1.871E+008 | 48478.158 | 1.870E+008 | 1.872E+008
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: Compiled) | Duration       | msec  |     27     |    377.521 |    20.081 |    335.727 |    419.947
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: None)     | Duration       | msec  |     15     |    709.439 |    31.986 |    655.717 |    765.097


   System.Text.RegularExpressions.Performance.Tests.dll                              | Metric         | Unit  | Iterations |    Average |   STDEV.S |        Min |        Max
  :--------------------------------------------------------------------------------- |:-------------- |:-----:|:----------:| ----------:| ---------:| ----------:| ----------:
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | Duration       | msec  |     45     |    226.408 |    13.236 |    212.146 |    262.963
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | GC Allocations | bytes |     45     | 1.871E+008 | 46358.265 | 1.869E+008 | 1.872E+008
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: Compiled) | Duration       | msec  |     28     |    366.970 |    17.217 |    337.323 |    407.299
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: None)     | Duration       | msec  |     15     |    700.781 |    35.247 |    653.294 |    790.779

## after
two runs:

  System.Text.RegularExpressions.Performance.Tests.dll                              | Metric         | Unit  | Iterations |    Average |   STDEV.S |        Min |        Max
  :--------------------------------------------------------------------------------- |:-------------- |:-----:|:----------:| ----------:| ---------:| ----------:| ----------:
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | Duration       | msec  |     46     |    220.359 |     7.715 |    213.496 |    247.767
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | GC Allocations | bytes |     46     | 1.871E+008 | 48166.588 | 1.870E+008 | 1.872E+008
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: Compiled) | Duration       | msec  |     29     |    345.858 |    22.669 |    318.472 |    412.110
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: None)     | Duration       | msec  |     15     |    667.597 |    33.483 |    624.545 |    745.706

  System.Text.RegularExpressions.Performance.Tests.dll                              | Metric         | Unit  | Iterations |    Average |   STDEV.S |        Min |        Max
  :--------------------------------------------------------------------------------- |:-------------- |:-----:|:----------:| ----------:| ---------:| ----------:| ----------:
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | Duration       | msec  |     45     |    226.161 |    15.460 |    212.438 |    300.492
   System.Text.RegularExpressions.Tests.Perf_Regex.Match                             | GC Allocations | bytes |     45     | 1.871E+008 | 59941.718 | 1.869E+008 | 1.872E+008
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: Compiled) | Duration       | msec  |     29     |    354.753 |    22.682 |    310.074 |    390.565
   System.Text.RegularExpressions.Tests.RegexRedux.RegexReduxMini(options: None)     | Duration       | msec  |     15     |    666.738 |    30.583 |    621.243 |    709.226 


The RegexReduxMini not compiled improves by 6% (a little better than 1 std dev).  Compiled improves by about 6% but this is within noise: I would only expect the division to help and then slightly. The regular Match case also remains the same within noise.

At this point I only made the change in the interpreted mode. I annotated the relevant compiled mode code which demonstrates it is not a simple transliteration into IL. That is future work. Meantime I believe the risk of this change diverging behavior is low because it is an extension of existing singleton behavior.