@danmosemsft 2018-04-05
@stephentoub with @ViktorHofer out, would you be able to review this? Should not need lots of context.