@danmosemsft 2018-04-06
We are going to defer this to 2.2 as it won't help published numbers for this benchmark (as that uses compiled) and then we'll be able to gather more data for real regexes (or current perf test has small inputs, and is not necessarily representative of real patterns)