@danmosemsft 2018-04-18
I'm going to close this, until I have time to look at the perf tests and the IL path.